requirements.txt
Django>=4.2.0
django-crispy-forms>=2.0.0
django-allauth>=0.54.0
Pillow>=9.5.0
django-environ>=0.10.0
psycopg2-binary>=2.9.6
djangorestframework>=3.14.0
python-dateutil>=2.8.2
django-tailwind>=3.6.0
whitenoise>=6.4.0
gunicorn>=20.1.0
